// web/src/api-client/models/ChatRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Session-Based Chatty LLM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * The request body for sending a new message to a chat session.
 * @export
 * @interface ChatRequest
 */
export interface ChatRequest {
    /**
     * 
     * @type {string}
     * @memberof ChatRequest
     */
    sessionId?: string | null;
    /**
     * The text content of the user's message.
     * @type {string}
     * @memberof ChatRequest
     */
    content: string;
}

/**
 * Check if a given object implements the ChatRequest interface.
 */
export function instanceOfChatRequest(value: object): value is ChatRequest {
    if (!('content' in value) || value['content'] === undefined) return false;
    return true;
}

export function ChatRequestFromJSON(json: any): ChatRequest {
    return ChatRequestFromJSONTyped(json, false);
}

export function ChatRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'sessionId': json['session_id'] == null ? undefined : json['session_id'],
        'content': json['content'],
    };
}

export function ChatRequestToJSON(json: any): ChatRequest {
    return ChatRequestToJSONTyped(json, false);
}

export function ChatRequestToJSONTyped(value?: ChatRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'session_id': value['sessionId'],
        'content': value['content'],
    };
}